Split old thread tests into new test module for multitasking module.
