+++ yggdrasil/tests/test_tools.py
@@ -238,47 +238,3 @@ class TestYggClass(YggTestClass):
             self.instance.sleep()
         assert(self.instance.check_timeout())
         self.instance.stop_timeout(quiet=True)
-
-
-class TestYggThread(YggTestClass):
-    r"""Test basic behavior of YggThread class."""
-
-    _cls = 'YggThread'
-    _mod = 'yggdrasil.tools'
-
-    def __init__(self, *args, **kwargs):
-        super(TestYggThread, self).__init__(*args, **kwargs)
-        self.namespace = 'TESTING_%s' % self.uuid
-        self.attr_list += ['name', 'sleeptime', 'longsleep', 'timeout']
-        self._inst_kwargs = {'timeout': self.timeout,
-                             'sleeptime': self.sleeptime}
-        self.debug_flag = False
-
-    def test_id(self):
-        r"""Test process ID and ident."""
-        self.instance.pid
-        self.instance.ident
-
-    def test_exitcode(self):
-        r"""Test process exitcode."""
-        self.instance.exitcode
-
-    def test_get_main_proc(self):
-        r"""Test get_main_proc."""
-        self.instance.get_main_proc()
-
-
-class TestYggProcess(TestYggThread):
-    r"""Test basic behavior of YggThread for spawned process."""
-
-    def __init__(self, *args, **kwargs):
-        super(TestYggProcess, self).__init__(*args, **kwargs)
-        self._inst_kwargs['method'] = 'process'
-
-
-class TestYggProcessFork(TestYggProcess):
-    r"""Test basic behavior of YggThread for forked process."""
-
-    def __init__(self, *args, **kwargs):
-        super(TestYggProcessFork, self).__init__(*args, **kwargs)
-        self._inst_kwargs['context'] = tools.mp_ctx
